refactor(sidebar): refactor to keep the main component as server component
