+++ src/app/(dashboard)/_sidebar/header.tsx
@@ -1,3 +1,5 @@
+'use client'
+
 import logo from '@images/logo/logo-triagem.svg'
 import { PanelLeftCloseIcon, PanelLeftOpenIcon } from 'lucide-react'
 import Image from 'next/image'
